refactor: 객체 테스트 분리
