+++ app/src/main/java/camp/nextstep/edu/tictactoe/MainViewModel.kt
@@ -5,26 +5,26 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class MainViewModel : ViewModel() {
-    private val ticTacToe: TicTacToe = TicTacToe()
+    private val ticTacToeBoard: TicTacToeBoard = TicTacToeBoard()
 
     private val _cellsLiveData: MutableLiveData<Array<Array<OX?>>> = MutableLiveData()
     val cellsLiveData: LiveData<Array<Array<OX?>>> get() = _cellsLiveData
 
-    private val _gameStatus = MutableLiveData(ticTacToe.getGameStatus())
+    private val _gameStatus = MutableLiveData(ticTacToeBoard.getGameStatus())
     val gameStatus: LiveData<TicTacToeStatus> get() = _gameStatus
 
     fun reset() {
-        ticTacToe.reset()
-        _cellsLiveData.value = ticTacToe.getAllCell()
-        _gameStatus.value = ticTacToe.getGameStatus()
+        ticTacToeBoard.reset()
+        _cellsLiveData.value = ticTacToeBoard.getAllCell()
+        _gameStatus.value = ticTacToeBoard.getGameStatus()
     }
 
     fun putCell(x: Int, y: Int) {
-        if (ticTacToe.getGameStatus() != TicTacToeStatus.PLAYING) return
+        if (ticTacToeBoard.getGameStatus() != TicTacToeStatus.PLAYING) return
 
-        ticTacToe.put(x, y)
-        _cellsLiveData.value = ticTacToe.getAllCell()
-        _gameStatus.value = ticTacToe.getGameStatus()
+        ticTacToeBoard.put(x, y)
+        _cellsLiveData.value = ticTacToeBoard.getAllCell()
+        _gameStatus.value = ticTacToeBoard.getGameStatus()
     }
 
 }
