+++ domain/src/main/java/camp/nextstep/edu/tictactoe/GameStatus.kt
@@ -1,10 +1,11 @@
 package camp.nextstep.edu.tictactoe
 
 
-internal class GameStatus(private val ticTacToe : Array<Array<OX?>>) {
+// 게임의 진행 상태
+internal class GameStatus(private val board : TicTacToeBoard) {
 
     val currentGameStatus: TicTacToeStatus
-        get() = getGameStatus(ticTacToe)
+        get() = getGameStatus()
 
     // 체크 순서
     var gameTern = OX.O
@@ -21,8 +22,8 @@ internal class GameStatus(private val ticTacToe : Array<Array<OX?>>) {
     // 배열이 가득 차거나 게임 승부가 나지 않은 상태
     private fun isDraw(isNotExistedWinner: Boolean): Boolean = isNotExistedWinner && (putCount == 9)
 
-    private fun getGameStatus(ticTacToe : Array<Array<OX?>>): TicTacToeStatus {
-        val winner = Winning.getWinner(ticTacToe)
+    private fun getGameStatus(): TicTacToeStatus {
+        val winner = Winning.getWinner(board)
         return when {
             isDraw(winner == null) -> TicTacToeStatus.DRAW
             winner == OX.X -> TicTacToeStatus.X_WIN
