+++ domain/src/main/java/camp/nextstep/edu/tictactoe/TicTacToeStatus.kt
@@ -1,5 +1,6 @@
 package camp.nextstep.edu.tictactoe
 
+// 게임의 현재 상태
 enum class TicTacToeStatus {
     O_WIN, X_WIN, DRAW, PLAYING
 }
